+++ docs/changelog.md
@@ -12,6 +12,17 @@ pageflow_next_text: Contributing
 
 Nothing yet.
 
+## v1.2016052401
+
+### Fixes
+
+* Skip the first frame of a debug stack trace - the information provided in there can never be complete
+  - Updated `FilterBacktrace`
+  - Updated `FilterBacktraceForTwoCodeCallers`
+* Make sure our base exceptions do not appear in the caller details that we include in an exception
+  - Updated `UnsupportedType`
+  - Updated `UnsupportedValue`
+
 ## v1.2016050201
 
 Released Mon 2nd May 2016.